Small change
Just changed some stuff- this seems better idk- also there is the bug of  it saying "extension reloaded" even if the cog does not exist lol. i dont think this throws an error but havent tested- if someone wants to throw it in a try catch to get easy com dev- go for it. once again havent tested- idk if this throw an error. but it isnt rly a problem. will make only jammy able to use reload soon.